fix: move integrations object mutation to the core SDK
