+++ packages/analytics-js-plugins/__tests__/xhrQueue/index.test.ts
@@ -309,7 +309,7 @@ describe('XhrQueue', () => {
           AnonymousId: 'c2FtcGxlQW5vbklk', // Base64 encoded anonymousId
         },
         sendRawData: true,
-        data: '{"batch":[{"type":"track","event":"test","userId":"test","properties":{"test":"test"},"anonymousId":"sampleAnonId","messageId":"test","originalTimestamp":"test","sentAt":"sample_timestamp","integrations":{"All":true}},{"type":"track","event":"test2","userId":"test2","properties":{"test2":"test2"},"anonymousId":"sampleAnonId","messageId":"test2","originalTimestamp":"test2","sentAt":"sample_timestamp","integrations":{"All":true}}]}',
+        data: '{"batch":[{"type":"track","event":"test","userId":"test","properties":{"test":"test"},"anonymousId":"sampleAnonId","messageId":"test","originalTimestamp":"test","sentAt":"sample_timestamp"},{"type":"track","event":"test2","userId":"test2","properties":{"test2":"test2"},"anonymousId":"sampleAnonId","messageId":"test2","originalTimestamp":"test2","sentAt":"sample_timestamp"}]}',
       },
       isRawResponse: true,
       timeout: 10000,
