+++ packages/analytics-js-plugins/src/utilities/queue.ts
@@ -1,13 +1,9 @@
 import { clone } from 'ramda';
 import { getCurrentTimeFormatted } from '@rudderstack/analytics-js-common/utilities/timestamp';
-import { mergeDeepRight } from '@rudderstack/analytics-js-common/utilities/object';
 import { stringifyWithoutCircular } from '@rudderstack/analytics-js-common/utilities/json';
 import { RudderEvent } from '@rudderstack/analytics-js-common/types/Event';
 import { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import { Nullable } from '@rudderstack/analytics-js-common/types/Nullable';
-import { IntegrationOpts } from '@rudderstack/analytics-js-common/types/Integration';
-import { ApplicationState } from '@rudderstack/analytics-js-common/types/ApplicationState';
-import { DEFAULT_INTEGRATIONS_CONFIG } from '@rudderstack/analytics-js-common/constants/integrationsConfig';
 import { EVENT_PAYLOAD_SIZE_BYTES_LIMIT } from './constants';
 import {
   EVENT_PAYLOAD_SIZE_CHECK_FAIL_WARNING,
@@ -52,49 +48,17 @@ const validateEventPayloadSize = (event: RudderEvent, logger?: ILogger) => {
 };
 
 /**
- * Filters and returns the user supplied integrations config that should take preference over the destination specific integrations config
- * @param eventIntgConfig User supplied integrations config at event level
- * @param destinationsIntgConfig Cumulative integrations config from all destinations
- * @returns Filtered user supplied integrations config
- */
-const getOverriddenIntegrationOptions = (
-  eventIntgConfig: IntegrationOpts,
-  destinationsIntgConfig: IntegrationOpts,
-): IntegrationOpts =>
-  Object.keys(eventIntgConfig)
-    .filter(intgName => eventIntgConfig[intgName] !== true || !destinationsIntgConfig[intgName])
-    .reduce((obj: IntegrationOpts, key: string) => {
-      const retVal = clone(obj);
-      retVal[key] = eventIntgConfig[key];
-      return retVal;
-    }, {});
-
-/**
  * Mutates the event and return final event for delivery
  * Updates certain parameters like sentAt timestamp, integrations config etc.
  * @param event RudderEvent object
  * @param state Application state
  * @returns Final event ready to be delivered
  */
-const getFinalEventForDeliveryMutator = (
-  event: RudderEvent,
-  state: ApplicationState,
-): RudderEvent => {
+const getFinalEventForDeliveryMutator = (event: RudderEvent): RudderEvent => {
   const finalEvent = clone(event);
 
   // Update sentAt timestamp to the latest timestamp
   finalEvent.sentAt = getCurrentTimeFormatted();
-
-  // Merge the destination specific integrations config with the event's integrations config
-  // In general, the preference is given to the event's integrations config
-  const eventIntgConfig = event.integrations ?? DEFAULT_INTEGRATIONS_CONFIG;
-  const destinationsIntgConfig = state.nativeDestinations.integrationsConfig.value;
-  const overriddenIntgOpts = getOverriddenIntegrationOptions(
-    eventIntgConfig,
-    destinationsIntgConfig,
-  );
-
-  finalEvent.integrations = mergeDeepRight(destinationsIntgConfig, overriddenIntgOpts);
 
   return finalEvent;
 };
@@ -102,7 +66,6 @@ const getFinalEventForDeliveryMutator = (
 export {
   getDeliveryPayload,
   validateEventPayloadSize,
-  getOverriddenIntegrationOptions,
   getFinalEventForDeliveryMutator,
   getBatchDeliveryPayload,
 };
